requirements.txt
openai-agents==0.0.7
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic>=2.10,<3.0 